move large scale validation function
