+++ data/users.js
@@ -1,38 +1,9 @@
 // Data functions for user profile objects
 
-import * as validation from "../utils/validation.js";
+import * as validation from "../public/js/validation.js";
 import { usersCollection } from "../config/mongoCollections.js";
-
-// Construct a user document given an object containing the required fields, or throw an error if any fields are invalid.
-// Set `allowUndefined` to `true` to ignore `undefined` values, i.e. create partial objects for PATCH requests.
-export function createUserDocument({ uid, password, firstName, lastName, description, profilePicture, availability }, allowUndefined = false) {
-    // ============= validate inputs =============
-
-    if (!allowUndefined || typeof uid !== "undefined") uid = validation.validateUserId(uid);
-    if (!allowUndefined || typeof password !== "undefined") password = validation.validateAndTrimString(password, "Password"); // should be hashed already
-    if (!allowUndefined || typeof firstName !== "undefined") firstName = validation.sanitizeSpaces(validation.validateAlphabeticalExtended(firstName, "First Name", 1));
-    if (!allowUndefined || typeof lastName !== "undefined") lastName = validation.sanitizeSpaces(validation.validateAlphabeticalExtended(lastName, "Last Name", 1));
-    if (!allowUndefined || typeof description !== "undefined") description = validation.validateAndTrimString(description, "Description", true);
-    if (!allowUndefined || typeof profilePicture !== "undefined") profilePicture = validation.validateAndTrimString(profilePicture, "Profile Picture");
-    if (!allowUndefined || typeof availability !== "undefined") availability = validation.validateArrayElements(availability, "Availability", (timeslot) => timeslot, 7); // todo - validate Timeslot objects
-
-    // ============= construct the document =============
-    const user = {
-        _id: uid,
-        password,
-        firstName,
-        lastName,
-        description,
-        profilePicture,
-        availability,
-    };
-
-    if (allowUndefined) {
-        // delete undefined properties from the final object
-        for (const key of Object.keys(user)) if (user[key] === undefined) delete user[key];
-    }
-    return user;
-}
+import { createUserDocument } from "../public/js/dbValidation.js";
+export { createUserDocument } from "../public/js/dbValidation.js";
 
 // create a user object and save it to the DB, then return the added object
 export async function createUser({ uid, password, firstName, lastName, description, profilePicture, availability }) {
